+++ src/components/product-card/product-menu.tsx
@@ -1,7 +1,7 @@
 import { RiAddFill, RiDeleteBinLine, RiSubtractFill } from 'react-icons/ri';
 import { messages } from '../../sources/messages';
 import { cartStore } from '../../store/cart-store';
-import { useCartHandlers } from '../../utils/hooks/cart-handlers';
+import { useCartHandlers } from '../../utils/hooks/use-cart-handlers';
 import { Button } from '../button/button';
 import styles from './product-card.module.css';
 
@@ -9,34 +9,24 @@ export const ProductMenu: React.FC<{ productId: string }> = ({ productId }) => {
   const isInCart = cartStore.isInCart(productId);
   const quantity = cartStore.getItemQuantity(productId);
   const { handleAddToCart, handleRemoveFromCart, handleUpdateQuantity } =
-    useCartHandlers();
+    useCartHandlers(productId);
   return (
     <>
       {isInCart ? (
         <div className={styles.itemMenu}>
-          <Button onClick={event => handleRemoveFromCart(productId, event)}>
+          <Button onClick={handleRemoveFromCart}>
             <RiDeleteBinLine />
           </Button>
-          <Button
-            onClick={event =>
-              handleUpdateQuantity(productId, quantity - 1, event)
-            }
-          >
+          <Button onClick={event => handleUpdateQuantity(quantity - 1, event)}>
             <RiSubtractFill />
           </Button>
           <span>{quantity}</span>
-          <Button
-            onClick={event =>
-              handleUpdateQuantity(productId, quantity + 1, event)
-            }
-          >
+          <Button onClick={event => handleUpdateQuantity(quantity + 1, event)}>
             <RiAddFill />
           </Button>
         </div>
       ) : (
-        <Button onClick={event => handleAddToCart(productId, event)}>
-          {messages.buttons.addToCart}
-        </Button>
+        <Button onClick={handleAddToCart}>{messages.buttons.addToCart}</Button>
       )}
     </>
   );
